chore: restructuring code
