+++ internal/validator/validator.go
@@ -2,12 +2,9 @@ package validator
 
 import (
 	"fmt"
-	"net/http"
 	"regexp"
 	"strconv"
 	"strings"
-
-	json "github.com/sugawarayuuta/sonnet"
 
 	"rudder-load/internal/parser"
 )
@@ -125,46 +122,6 @@ func ValidateLoadTestConfig(config *parser.LoadTestConfig) error {
 		}
 		if phase.Replicas <= 0 {
 			return fmt.Errorf("replicas must be greater than 0")
-		}
-	}
-	return nil
-}
-
-// Define the expected JSON structure to parse
-type userTransformerHashEmailResponse []struct {
-	Output struct {
-		Context struct {
-			Traits struct {
-				Email string `json:"email"`
-			} `json:"traits"`
-		} `json:"context"`
-	} `json:"output"`
-	StatusCode int `json:"statusCode"`
-}
-
-func ValidateResponseBody(validatorType string) func(body []byte) (bool, error) {
-	switch validatorType {
-	case "user-transformer-hash-email":
-		return func(body []byte) (bool, error) {
-			var parsedBody userTransformerHashEmailResponse
-			if err := json.Unmarshal(body, &parsedBody); err != nil {
-				return false, fmt.Errorf("invalid response body JSON: %w", err)
-			}
-			if len(parsedBody) == 0 {
-				return false, fmt.Errorf("response body array is empty")
-			}
-			for _, item := range parsedBody {
-				if item.StatusCode != http.StatusOK {
-					return false, fmt.Errorf("response status code is not 200")
-				}
-				if item.Output.Context.Traits.Email == "" {
-					return false, fmt.Errorf("email trait is missing")
-				}
-				if !sha256Validator.MatchString(item.Output.Context.Traits.Email) {
-					return false, fmt.Errorf("email hash must be a valid sha256 hexadecimal string")
-				}
-			}
-			return true, nil
 		}
 	}
 	return nil
